Explicitly call toInt() on this QVariant
I'm not sure why this isn't working for me out of the box.  Maybe the
type conversions must be precise?

In my case, `setSize()` expects either `(QSize)` or `(int, int)`.  `height`
is a `QVariant` containing a `float`, and this is causing my pyqt to
rais an unhandled exception.  Explicitly call toInt() to get qt to use
the proper signature when invoking setSize.

+++ fbmessengerlib/windows.py
@@ -90,7 +90,7 @@ def _terrible_toast_height_hack():
     height = toast_window.evaluate_js(
         'document.getElementById("toast-frame").offsetHeight')
     if height:
-      toast_window.set_size(TOAST_WIDTH, height)
+      toast_window.set_size(TOAST_WIDTH, height.toInt())
       _position_toast()
     else:
       print("Failed to hack out the toast height.")
